src/importer/schema_st4/requirements.txt
schema-st4-parser
pandas
farm-haystack
